refactor(nextcloud): separate webdav utils and models from the webdav/client library
