+++ packages/nextcloud/lib/webdav.dart
@@ -4,8 +4,10 @@ import 'package:nextcloud/src/webdav/client.dart';
 
 export 'src/webdav/client.dart' hide WebDavRequest;
 export 'src/webdav/file.dart';
+export 'src/webdav/models.dart';
 export 'src/webdav/path_uri.dart';
 export 'src/webdav/props.dart';
+export 'src/webdav/utils.dart' hide constructUri;
 export 'src/webdav/webdav.dart';
 
 // ignore: public_member_api_docs
